Modified Rocket_Flight_DM.cpp
FSPB is the XCG acceleration can be sensed by IMU, but it is in body frame,
and the gerenal dynamics euqation is to calculate acceleration at the
reference point in inertia frame, so that need to compensate by the ddrhoC_IMU or ddrhoC_1,
these parameter is in body frame.

But the GRAVG, NEXT_ACC is in inertia frame, so that ddrhoC_1 or
ddrhoC_IMU need be rotated to the inertia frame.